Its working - without unit tests
